# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.ChangePasswordInput:
    properties:
      new_password:
        example: newpassword
        maxLength: 72
        minLength: 8
        type: string
      old_password:
        example: password
        maxLength: 72
        minLength: 8
        type: string
    required:
    - new_password
    - old_password
    type: object
  domain.ChangeRoleInput:
    properties:
      role:
        enum:
        - admin
        - member
        example: member
        type: string
    required:
    - role
    type: object
...
  domain.RegisterInput:
    properties:
      email:
        example: user@example.com
        maxLength: 254
        type: string
      password:
        example: password
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: error description
        type: string
    type: object
...
  handler.UserResponse:
    properties:
      created_at:
        example: "2025-01-15T09:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      id:
        example: 7bc4f9d2-1a3e-4c8f-9b2d-5e7a0c4f1d63
        type: string
      role:
        example: member
        type: string
      updated_at:
        example: "2025-01-15T09:00:00Z"
        type: string
    type: object
host: 204.168.173.88:8080
info:
  contact:
    name: Mykyta Kuzminov
    url: https://github.com/mykytakuzminov
  description: |-
    REST API for managing boards, columns and tasks with JWT authentication.
    Supports token-based auth, rate limiting, pagination, sorting and filtering.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Sprintly API
  version: "1.0"
paths:
  /admin/users/{id}:
    delete:
      description: Deletes a user by ID. Requires admin role.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: User deleted successfully
        "400":
          description: Invalid user ID format
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: admin role required'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user
      tags:
      - admin
    get:
      description: Returns the profile of a specific user by their ID. Requires admin
        role.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: admin role required'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - admin
    patch:
      consumes:
      - application/json
      description: Changes the role of the given user. Requires admin role.
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: new role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.ChangeRoleInput'
      produces:
      - application/json
      responses:
        "204":
          description: Role changed successfully
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: admin role required'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change role
      tags:
      - admin
  /admin/users/all:
    get:
      description: Returns all users. Requires admin role. Supports pagination and
        sorting.
      parameters:
      - description: 'Number of results per page (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      - description: 'Number of results to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - description: 'Field to sort by: email, created_at, updated_at (default: created_at)'
        in: query
        name: sort
        type: string
      - description: 'Sort direction: ASC or DESC (default: ASC)'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            items:
              $ref: '#/definitions/handler.UserResponse'
            type: array
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: admin role required'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - admin
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account. Email must be unique.
      parameters:
      - description: Registration data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
  /users/me:
    get:
      description: Returns the profile of the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Current user data
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "401":
          description: Missing or invalid access token
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
  /users/me/password:
    patch:
      consumes:
      - application/json
      description: Changes the password of the currently authenticated user. Requires
        the current password for verification.
      parameters:
      - description: Old and new passwords
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.ChangePasswordInput'
      produces:
      - application/json
      responses:
        "204":
          description: Password changed successfully
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Missing or invalid access token or wrong old password
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change password
      tags:
      - users
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your Bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"